requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.26.0
qrcode[pil]==7.4.2
Pillow==10.2.0
requests==2.31.0
cryptography==42.0.2
pydantic[email]==2.6.1
bakong-khqr
